# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PlusMove Laravel API",
        "description": "API documentation for PlusMove which handles deliveries.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/user/{id}/address": {
            "post": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Store a new address for a user",
                "operationId": "a643b1cc13cf2891477074cbaa1ad77c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user to store a new for the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 20
                                                },
                                                "full_name": {
                                                    "type": "string",
                                                    "example": "Thomas Opperman"
                                                },
                                                "phone_number": {
                                                    "type": "string",
                                                    "example": "0574911775"
                                                },
                                                "role": {
                                                    "type": "string",
                                                    "example": "Admin"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "format": "email",
                                                    "example": "mark.grobbelaar@example.com"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "1 day ago"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "1 day ago"
                                                },
                                                "addresses": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 39
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Home"
                                                            },
                                                            "address_line_1": {
                                                                "type": "string",
                                                                "example": "2839 Village Port"
                                                            },
                                                            "address_line_2": {
                                                                "type": "string",
                                                                "example": "5362 Sandile Skyway Drive",
                                                                "nullable": true
                                                            },
                                                            "suburb": {
                                                                "type": "string",
                                                                "example": "Klarinet"
                                                            },
                                                            "city": {
                                                                "type": "string",
                                                                "example": "Pretoria"
                                                            },
                                                            "province": {
                                                                "type": "string",
                                                                "example": "North West"
                                                            },
                                                            "postal_code": {
                                                                "type": "string",
                                                                "example": "1358"
                                                            },
                                                            "country": {
                                                                "type": "string",
                                                                "example": "South Africa"
                                                            },
                                                            "created_at": {
                                                                "type": "string",
                                                                "example": "1 day ago"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The address line 1 field is required."
                                        },
                                        "errors": {
                                            "properties": {
                                                "address_line_1": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The address line 1 field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - The user does not have permission to view the addresses for this user."
                    },
                    "404": {
                        "description": "Not Found - The user with the specified ID was not found."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/customer/{id}/address": {
            "post": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Store a new address for a customer",
                "description": "Store a new address for a customer",
                "operationId": "8973bf79cfde8fb9f831efaee4fcc160",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the customer to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Miss Willemse"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "012-712-3489"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "format": "email",
                                                    "example": "norman79@example.org"
                                                },
                                                "joined": {
                                                    "type": "string",
                                                    "example": "1 day ago"
                                                },
                                                "customer_addresses": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 71
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Home"
                                                            },
                                                            "address_line_1": {
                                                                "type": "string",
                                                                "example": "20288 Nomvula Meadows"
                                                            },
                                                            "address_line_2": {
                                                                "type": "string",
                                                                "example": "8042 Gary Ridge Road",
                                                                "nullable": true
                                                            },
                                                            "suburb": {
                                                                "type": "string",
                                                                "example": "Roodepan"
                                                            },
                                                            "city": {
                                                                "type": "string",
                                                                "example": "Pietermaritzburg"
                                                            },
                                                            "province": {
                                                                "type": "string",
                                                                "example": "Northern Cape"
                                                            },
                                                            "postal_code": {
                                                                "type": "string",
                                                                "example": "3539"
                                                            },
                                                            "country": {
                                                                "type": "string",
                                                                "example": "South Africa"
                                                            },
                                                            "created_at": {
                                                                "type": "string",
                                                                "example": "1 day ago"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - The user does not have permission to view this customer."
                    },
                    "404": {
                        "description": "Not Found - The customer with the specified ID was not found."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}